+++ ESFramework/Foundation/NSArray+ESAdditions.h
@@ -8,13 +8,20 @@
 
 #import <Foundation/Foundation.h>
 
+NS_ASSUME_NONNULL_BEGIN
+
 @interface NSArray<ObjectType> (ESAdditions)
 
-- (ObjectType)objectPassingTest:(BOOL (^)(ObjectType obj, NSUInteger idx, BOOL *stop))predicate;
-- (ObjectType)objectWithOptions:(NSEnumerationOptions)opts passingTest:(BOOL (^)(ObjectType obj, NSUInteger idx, BOOL *stop))predicate;
+- (nullable ObjectType)objectPassingTest:(BOOL (^)(ObjectType obj, NSUInteger idx, BOOL *stop))predicate;
+- (nullable ObjectType)objectWithOptions:(NSEnumerationOptions)opts passingTest:(BOOL (^)(ObjectType obj, NSUInteger idx, BOOL *stop))predicate;
 
 - (NSArray<ObjectType> *)objectsPassingTest:(BOOL (^)(ObjectType obj, NSUInteger idx, BOOL *stop))predicate;
 - (NSArray<ObjectType> *)objectsWithOptions:(NSEnumerationOptions)opts passingTest:(BOOL (^)(ObjectType obj, NSUInteger idx, BOOL *stop))predicate;
+
+- (nullable ObjectType)previousObjectToIndex:(NSUInteger)index;
+- (nullable ObjectType)previousObjectToObject:(ObjectType)object;
+- (nullable ObjectType)nextObjectToIndex:(NSUInteger)index;
+- (nullable ObjectType)nextObjectToObject:(ObjectType)object;
 
 @end
 
@@ -23,3 +30,5 @@
 - (void)replaceObject:(ObjectType)object withObject:(ObjectType)anObject;
 
 @end
+
+NS_ASSUME_NONNULL_END